[feature](Nereids): eager aggreagate support mix agg function
