+++ regression-test/data/nereids_rules_p0/eager_aggregate/push_down_min_through_join.out
@@ -148,8 +148,10 @@ PhysicalResultSink
 --hashAgg[GLOBAL]
 ----hashAgg[LOCAL]
 ------hashJoin[INNER_JOIN] hashCondition=((t1.id = t2.id) and (t1.name = t2.name)) otherCondition=()
---------PhysicalOlapScan[min_t]
---------PhysicalOlapScan[min_t]
+--------hashAgg[LOCAL]
+----------PhysicalOlapScan[min_t]
+--------hashAgg[LOCAL]
+----------PhysicalOlapScan[min_t]
 
 -- !groupby_pushdown_equal_conditions_non_aggregate_selection --
 PhysicalResultSink
