+++ regression-test/suites/nereids_rules_p0/eager_aggregate/basic_one_side.groovy
@@ -21,9 +21,7 @@ suite("eager_aggregate_basic_one_side") {
     sql "SET enable_fallback_to_original_planner=false"
     sql "SET ignore_shape_nodes='PhysicalDistribute,PhysicalProject'"
 
-    sql "SET ENABLE_NEREIDS_RULES=push_down_min_max_through_join_one_side"
-    sql "SET ENABLE_NEREIDS_RULES=push_down_sum_through_join_one_side"
-    sql "SET ENABLE_NEREIDS_RULES=push_down_count_through_join_one_side"
+    sql "SET ENABLE_NEREIDS_RULES=push_down_agg_through_join_one_side"
 
     sql """
         DROP TABLE IF EXISTS shunt_log_com_dd_library_one_side;
