+++ regression-test/suites/nereids_rules_p0/eager_aggregate/push_down_max_through_join.groovy
@@ -48,7 +48,7 @@ suite("push_down_max_through_join") {
     sql "insert into max_t values (9, 3, null)"
     sql "insert into max_t values (10, null, null)"
 
-    sql "SET ENABLE_NEREIDS_RULES=push_down_min_max_through_join"
+    sql "SET ENABLE_NEREIDS_RULES=push_down_agg_through_join_one_side"
 
     qt_groupby_pushdown_basic """
         explain shape plan select max(t1.score) from max_t t1, max_t t2 where t1.id = t2.id group by t1.name;
